requirements.txt
python-dotenv==1.0.1
sentry-sdk==2.0.0
